backend/consume_data/requirements.txt
bitcoinaverage==0.3.1
blinker==1.9.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
dnspython==2.8.0
Flask==3.1.2
flask-cors==6.0.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mongoengine==0.29.1
path==17.1.1
pymongo==4.14.1
python-dotenv==1.1.1
requests==2.32.5
six==1.17.0
urllib3==2.5.0
Werkzeug==3.1.3
ws4py==0.6.0
